cat: Refactor splice code, handle more failures
